[refactor] add lottie to error handing
