+++ package.json
@@ -24,8 +24,10 @@
         "@twa-dev/sdk": "^8.0.1",
         "dayjs": "^1.11.13",
         "express": "^4.19.2",
+        "lottie-react": "^2.4.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
+        "react-lottie": "^1.2.10",
         "react-redux": "^9.2.0",
         "react-router-dom": "^6.28.0",
         "react-select": "^5.8.3"
@@ -42,4 +44,3 @@
         "typescript-eslint": "^8.13.0"
     }
 }
-
