+++ src/container/ErrorPage/icon/index.style.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const IconErrorStyled = styled.img`
+  width: 100%;
+  hight: 100%;
+`;
