+++ src/container/ErrorPage/info/index.style.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const InfoStyled = styled.p`
+  text-align: center;
+  color: var(--tg-theme-text-color);
+`;
