+++ src/container/ErrorPage/title/index.style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const TitleStyled = styled.h1`
+  font-size: 28px;
+  text-align: center;
+  color: var(--tg-theme-button-color);
+`;
