--- src/container/error/icon/index.style.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const IconErrorStyled = styled.img`
-  width: 100%;
-  hight: 100%;
-`;
