--- src/container/error/info/index.style.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const InfoStyled = styled.p`
-  text-align: center;
-  color: var(--tg-theme-text-color);
-`;
