--- src/container/error/info/index.tsx
@@ -1,8 +1,0 @@
-import React, { ReactNode } from 'react';
-import { InfoStyled } from './index.style';
-
-const Info = ({ children }: { children: ReactNode }): React.ReactElement => {
-  return <InfoStyled>{children}</InfoStyled>;
-};
-
-export default Info;
