+++ src/container/main/router.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 import SingUpPage from '../user/signup';
 import EditUserPage from '../user/edit';
-import ErrorPage from '../error';
-import { ErrorPageCode } from '../error/types';
+import ErrorPage from '../ErrorPage';
+import { ErrorPageCode } from '../ErrorPage/types';
 import { getNavigationsValue } from '@brojs/cli';
 import Layout from './components/layout';
 import EventList from '../event/EventList';
